# docs/swagger.yaml
definitions:
  papi.agency:
    properties:
      coletando:
        items:
          $ref: '#/definitions/papi.collecting'
        type: array
      entidade:
        description: '"J" For Judiciário, "M" for Ministério Público, "P" for Procuradorias
          and "D" for Defensorias.'
        type: string
      id_orgao:
        description: '''trt13'''
        type: string
      jurisdicao:
        description: '"R" for Regional, "M" for Municipal, "F" for Federal, "E" for
          State.'
        type: string
      nome:
        description: '''Tribunal Regional do Trabalho 13° Região'''
        type: string
      ouvidoria:
        description: Agencys's ombudsman url
        type: string
      twitter_handle:
        description: Agency's twitter handle
        type: string
      uf:
        description: Short code for federative unity.
        type: string
      url:
        description: Link for state url
        type: string
    type: object
  papi.collecting:
    properties:
      descricao:
        description: Reasons why we didn't collect the data
        items:
          type: string
        type: array
      timestamp:
        description: Day(unix) we checked the status of the data
        type: integer
    type: object
  uiapi.agencySalary:
    properties:
      histograma:
        additionalProperties:
          type: integer
        type: object
      max_salario:
        type: number
      package:
        $ref: '#/definitions/uiapi.backup'
    type: object
  uiapi.backup:
    properties:
      hash:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
  uiapi.procInfo:
    properties:
      cmd:
        type: string
      cmd_dir:
        type: string
      env:
        items:
          type: string
        type: array
      status:
        type: integer
      stderr:
        type: string
      stdin:
        type: string
      stdout:
        type: string
    type: object
  uiapi.timestamp:
    properties:
      nanos:
        type: integer
      seconds:
        type: integer
    type: object
  uiapi.v2AgencySummary:
    properties:
      max_outras_remuneracoes:
        type: number
      max_remuneracao_base:
        type: number
      orgao:
        type: string
      outras_remuneracoes:
        type: number
      remuneracao_base:
        type: number
      tem_anterior:
        type: boolean
      tem_proximo:
        type: boolean
      timestamp:
        $ref: '#/definitions/uiapi.timestamp'
      total_membros:
        type: integer
      total_remuneracao:
        type: number
    type: object
  uiapi.v2ProcInfoResult:
    properties:
      proc_info:
        $ref: '#/definitions/uiapi.procInfo'
      timestamp:
        $ref: '#/definitions/uiapi.timestamp'
    type: object
info:
  contact:
    name: DadosJusBr
    url: https://dadosjusbr.org
  title: API do dadosjusbr.org
  version: "1.0"
paths:
  /uiapi/v2/orgao/resumo/{orgao}/{ano}/{mes}:
    get:
      description: Resume os dados de remuneração mensal de um órgão.
      operationId: GetSummaryOfAgency
      parameters:
      - description: 'ID do órgão. Exemplos: tjal, tjba, mppb.'
        in: path
        name: orgao
        required: true
        type: string
      - description: 'Ano da remuneração. Exemplo: 2018.'
        in: path
        name: ano
        required: true
        type: integer
      - description: 'Mês da remuneração. Exemplo: 1.'
        in: path
        name: mes
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Requisição bem sucedida.
          schema:
            $ref: '#/definitions/uiapi.v2AgencySummary'
        "400":
          description: Parâmetro ano, mês ou nome do órgão são inválidos.
          schema:
            type: string
        "404":
          description: Órgão não encontrado.
          schema:
            type: string
      tags:
      - ui_api
  /uiapi/v2/orgao/salario/{orgao}/{ano}/{mes}:
    get:
      description: Busca dados das remunerações mensais de um órgão.
      operationId: GetSalaryOfAgencyMonthYear
      parameters:
      - description: 'ID do órgão. Exemplos: tjal, tjba, mppb.'
        in: path
        name: orgao
        required: true
        type: string
      - description: 'Mês da remuneração. Exemplos: 01, 02, 03...'
        in: path
        name: mes
        required: true
        type: string
      - description: 'Ano da remuneração. Exemplos: 2018, 2019, 2020...'
        in: path
        name: ano
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Requisição bem sucedida.
          schema:
            $ref: '#/definitions/uiapi.agencySalary'
        "206":
          description: Requisição bem sucedida, mas os dados do órgão não foram bem
            processados
          schema:
            $ref: '#/definitions/uiapi.v2ProcInfoResult'
        "400":
          description: Parâmetros inválidos.
          schema:
            type: string
      tags:
      - ui_api
  /v1/orgao/{orgao}:
    get:
      description: Busca um órgão específico utilizando seu ID.
      operationId: GetAgencyById
      parameters:
      - description: 'ID do órgão. Exemplos: tjal, tjba, mppb.'
        in: path
        name: orgao
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Requisição bem sucedida.
          schema:
            $ref: '#/definitions/papi.agency'
        "404":
          description: Órgão não encontrado.
          schema:
            type: string
      tags:
      - public_api
swagger: "2.0"